catalog_ops/requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
factory_boy==3.3.3
Faker==37.4.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
